Fix dependent tests for recent unit testing refactorings.
