+++ test/unit/test_model_copy.py
@@ -7,7 +7,7 @@ import galaxy.model
 import galaxy.model.mapping as mapping
 from galaxy.model.metadata import MetadataTempFile
 from galaxy.util import ExecutionTimer
-from .test_objectstore import DISK_TEST_CONFIG, TestConfig
+from .objectstore.test_objectstore import DISK_TEST_CONFIG, TestConfig
 
 
 datatypes_registry = galaxy.datatypes.registry.Registry()
